+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 1.3.2
+### Added
+* enableFetchLimitOverflowSoftNormalizer (default is false) to CoreDataCacheTracker
+
+### Changed
+* Suppressed some compilation warnings
+* Made demo project compatible with swift 4.2
+
 ## 1.3.1
 ### Added
 * fetchLimitThreshold